Default to dual stack and add BIND arg
Change the default bind address to be `[::]` and allow the user to
specify the socket to bind on.

This is distinct from the HOST/PORT pair already configurable as phd
might be running behind a proxy where links need to be generated with a
different destination to the bound socket.